修改中，新增GpxSerializer，拆分GpxSegment
